ctdb: Fix format errors for time_t!=long

+++ ctdb/server/ctdb_daemon.c
@@ -1134,7 +1134,7 @@ static void ctdb_tevent_trace(enum tevent_trace_point tp,
 		if (diff.tv_sec > 3) {
 			DEBUG(DEBUG_ERR,
 			      ("Handling event took %ld seconds!\n",
-			       diff.tv_sec));
+			       (long)diff.tv_sec));
 		}
 		tevent_before_wait_ts = now;
 		break;
@@ -1144,7 +1144,7 @@ static void ctdb_tevent_trace(enum tevent_trace_point tp,
 		if (diff.tv_sec > 3) {
 			DEBUG(DEBUG_ERR,
 			      ("No event for %ld seconds!\n",
-			       diff.tv_sec));
+			       (long)diff.tv_sec));
 		}
 		tevent_after_wait_ts = now;
 		break;
